[rbp/omxplayer] Remove discontinuity handling from players
I think sync/continuity checking done in OMXPlayerVideo and OMXPlayerAudio on Pi
is always wrong and leads to audio sync errors after discontinuities.

The GPU already does clock adjustment based on discontinuities in timestamps,
and is in a better position to do this adjustment.
(Remember there are a couple of seconds of buffered audio/video between here
and GPU and we want to adjust the clock at the other end of this queue).