[rbp/omxplayer] Make audio_render configurable as the clock master
By default audio_render is the clock master, and if output samples don't fit the timestamps, it will speed up/slow down the clock.
This tends to be better for maintaining audio sync and avoiding audio glitches, but can affect video/display sync.
The other option makes audio_render the clock slave, and it has to drop packet or output silence to maintain sync.

As both settings have some advantages, and as this mode more accurately defines whether playback is synced to display or audio, I've used the GUI setting "sync playback to display" to control this.

The previous use of "sync playback to display" didn't have any benefits, and was never recommended.