[OverlayRenderGL] Fix corrupt pixels in graphical subs
On Pi you can see a corrupt line of pixels on bottom and right edges of subtitle bounding box.
This is due to using a subregion of a texture (due to the power of two rounding) and not ensuring
the texture coordinates are inside the valid pixels.

The easy solution is to use NPOT when available (always on GLES) and then the clamping works
(as well as saving memory).